+++ cmd/template/cluster/runner.go
@@ -64,6 +64,7 @@ func (r *runner) run(ctx context.Context, cmd *cobra.Command, args []string) err
 			MasterAZ:       r.flag.MasterAZ,
 			Description:    r.flag.Name,
 			Owner:          r.flag.Owner,
+			PodsCIDR:       r.flag.PodsCIDR,
 			ReleaseVersion: r.flag.Release,
 			Namespace:      metav1.NamespaceDefault,
 		}