+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project's packages adheres to [Semantic Versioning](http://semver.org/s
 ### Changed
 
 - Make the `login` command validate the current context before considering it good to go.
-- Compute pod CIDR flag for AWS provider.
+- Re-enabled the `--pods-cidr` flag in the `template cluster` command.
 
 ## [1.20.0] - 2021-01-18