+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project's packages adheres to [Semantic Versioning](http://semver.org/s
 
 ## [Unreleased]
 
+## [1.21.0] - 2021-01-29
+
 ### Changed
 
 - Make the `login` command validate the current context before considering it good to go.
@@ -257,7 +259,8 @@ This release supports rendering for CRs:
 - `AppCatalog`
 - `App`
 
-[Unreleased]: https://github.com/giantswarm/kubectl-gs/compare/v1.20.0...HEAD
+[Unreleased]: https://github.com/giantswarm/kubectl-gs/compare/v1.21.0...HEAD
+[1.21.0]: https://github.com/giantswarm/kubectl-gs/compare/v1.20.0...v1.21.0
 [1.20.0]: https://github.com/giantswarm/kubectl-gs/compare/v1.19.0...v1.20.0
 [1.19.0]: https://github.com/giantswarm/kubectl-gs/compare/v0.18.0...v1.19.0
 [0.18.0]: https://github.com/giantswarm/kubectl-gs/compare/v0.17.0...v0.18.0